[model symptm frequency] - fixed the issue about rounding.

+++ sears-partsdirect-common/src/main/java/com/spd/cq/searspartsdirect/common/model/spdasset/SymptomModel.java
@@ -43,7 +43,7 @@ public class SymptomModel {
 		this.path = path;
 	}
 
-	public double getFrequency() {
+	public long getFrequency() {
 		return frequency;
 	}
